Refactor file structure and make basic all users page for admin
