--- frontend/SW-Quality-Dashboard/src/pages/AdminAllProjectsPage/AdminAllProjectsPage.css
@@ -1,18 +1,0 @@
-.dashboard-pill {
-    padding: 0.5em 1em;
-    border-radius: 5em;
-    color: white;
-    font-weight: bold;
-    font-size: 0.8em;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    text-align: center;
-    width: 70%;
-}
-
-.status {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
