--- frontend/SW-Quality-Dashboard/src/pages/AdminPage/AdminPage.tsx
@@ -1,18 +1,0 @@
-import './AdminPage.css'
-
-function AdminPage({icon, title, children}) {
-    return (
-        <div className="admin-page">
-            <h2>
-                <i className={icon}> </i>
-                {title}
-            </h2>
-
-            <div className="admin-page-content">
-                {children}
-            </div>
-        </div>
-    )
-}
-
-export default AdminPage
