+++ frontend/SW-Quality-Dashboard/src/pages/admin/AdminAllProjectsPage/AdminAllProjectsPage.css
@@ -1,0 +1,20 @@
+.dashboard-pill {
+    padding: 0.5em 1em;
+    border-radius: 5em;
+    color: white;
+    font-weight: bold;
+    font-size: 0.8em;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    text-align: center;
+    width: 70%;
+    box-shadow: rgba(50, 50, 93, 0.25) 0px 50px 100px -20px, rgba(0, 0, 0, 0.3) 0px 30px 60px -30px, rgba(10, 37, 64, 0.35) 0px -2px 6px 0px inset;
+
+}
+
+.status {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
