+++ frontend/SW-Quality-Dashboard/src/pages/admin/AdminTabContent/AdminTabContent.css
