+++ frontend/SW-Quality-Dashboard/src/pages/auth/ForgotPasswordPage/ForgotPasswordPage.css
