+++ frontend/SW-Quality-Dashboard/src/pages/auth/LoginPage/LoginPage.css
