+++ frontend/SW-Quality-Dashboard/src/pages/auth/RegisterPage/RegisterPage.css
