+++ frontend/SW-Quality-Dashboard/src/pages/auth/VerifyUserPage/VerifyUserPage.css
