+++ frontend/SW-Quality-Dashboard/src/pages/general/AboutPage/AboutPage.tsx
