+++ frontend/SW-Quality-Dashboard/src/pages/general/DashboardPage/DashboardPage.css
