+++ frontend/SW-Quality-Dashboard/src/pages/general/HomePage/HomePage.css
