+++ frontend/SW-Quality-Dashboard/src/pages/general/SubmitProjectPage/SubmitProjectPage.css
